requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
azure-identity==1.20.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cobble==0.1.4
cryptography==44.0.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.10
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.8.2
lxml==5.3.1
mammoth==1.9.0
markdownify==1.0.0
markitdown==0.0.1a5
msal==1.31.1
msal-extensions==1.2.0
numpy==2.2.3
olefile==0.47
openai==1.65.1
openpyxl==3.1.5
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20240706
pillow==11.1.0
portalocker==2.10.1
puremagic==1.28
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
starlette==0.46.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xlrd==2.0.1
XlsxWriter==3.2.2
youtube-transcript-api==0.6.3
